requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
Django==3.2.7
future==0.18.2
idna==3.2
joblib==1.0.1
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-whois==0.7.3
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.2.0
urllib3==1.26.6
